updated app design with download buttons
