+++ src/components/layout/Maincontent.vue
@@ -1,0 +1,5 @@
+<template>
+    <v-container class="fill-height d-flex align-center justify-center">
+        <slot /> <!-- Content will be injected here -->
+    </v-container>
+</template>
