--- src/pages/index.vue
@@ -1,7 +1,0 @@
-<template>
-  <HelloWorld />
-</template>
-
-<script setup>
-  //
-</script>
