+++ src/router/index.js
@@ -7,7 +7,12 @@
 // Composables
 import { createRouter, createWebHistory } from 'vue-router/auto'
 import { setupLayouts } from 'virtual:generated-layouts'
-import { routes } from 'vue-router/auto-routes'
+// import { routes } from 'vue-router/auto-routes'
+import HomePage from "@/pages/HomePage.vue";
+
+const routes = [
+  { path: "/", component: HomePage }
+];
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -34,3 +39,4 @@ router.isReady().then(() => {
 })
 
 export default router
+
